Define computer operations as separate functions

+++ intcode/list.h
@@ -1,13 +1,13 @@
 typedef struct {
-	int *values;
+	long *values;
 	int length;
 	int size;
 	int shift;
 } list_t;
 
 list_t init_list();
-int at(list_t*, int);
+long at(list_t*, int);
 void print(list_t*, char*);
-void append(list_t*, int);
-void set(list_t*, int, int);
-int shift(list_t*);
+void append(list_t*, long);
+void set(list_t*, int, long);
+long shift(list_t*);
